Fix repo source migration and reuse module reader
